requirements.txt
requests~=2.26.0
python-dotenv~=0.19.2
Flask~=2.0.2
WTForms~=3.0.0
SQLAlchemy~=1.4.27